+++ linkerd/core/src/main/scala/io/buoyant/linkerd/ProtocolInitializer.scala
@@ -54,7 +54,8 @@ abstract class ProtocolInitializer extends ConfigInitializer { initializer =>
    */
   private case class ProtocolRouter(
     router: StackRouter[RouterReq, RouterRsp],
-    servers: Seq[Server] = Nil
+    servers: Seq[Server] = Nil,
+    announcers: Seq[(Path, Announcer)] = Nil
   ) extends Router {
     def params = router.params
     def protocol = ProtocolInitializer.this
@@ -67,6 +68,8 @@ abstract class ProtocolInitializer extends ConfigInitializer { initializer =>
 
     protected def withServers(ss: Seq[Server]): Router = copy(servers = ss)
 
+    def withAnnouncers(ann: Seq[(Path, Announcer)]): Router = copy(announcers = ann)
+
     def initialize(): Router.Initialized = {
       if (servers.isEmpty) {
         val Label(name) = params[Label]
@@ -77,9 +80,9 @@ abstract class ProtocolInitializer extends ConfigInitializer { initializer =>
       val adapted = adapter.andThen(factory)
       val servable = servers.map { server =>
         val stackServer = defaultServer.withParams(defaultServer.params ++ server.params)
-        ServerInitializer(protocol, server.addr, stackServer, adapted)
+        ServerInitializer(protocol, server.addr, stackServer, adapted, server.announce)
       }
-      InitializedRouter(protocol, params, factory, servable)
+      InitializedRouter(protocol, params, factory, servable, announcers)
     }
 
     private[this] val tlsPrepRole = Stack.Role("TlsClientPrep")
@@ -130,7 +133,8 @@ object ProtocolInitializer {
     protocol: ProtocolInitializer,
     params: Stack.Params,
     factory: ServiceFactory[Req, Rsp],
-    servers: Seq[Server.Initializer]
+    servers: Seq[Server.Initializer],
+    announcers: Seq[(Path, Announcer)]
   ) extends Router.Initialized {
     def name: String = params[Label].label
     def close(t: Time) = factory.close(t)
@@ -141,7 +145,8 @@ object ProtocolInitializer {
     protocol: ProtocolInitializer,
     addr: InetSocketAddress,
     server: StackServer[Req, Rsp],
-    factory: ServiceFactory[Req, Rsp]
+    factory: ServiceFactory[Req, Rsp],
+    announce: Seq[Path]
   ) extends Server.Initializer {
     def params = server.params
     def router: String = server.params[Server.RouterLabel].label