+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Router.scala
@@ -62,6 +62,8 @@ trait Router {
   /** Return a router with TLS configuration read from the provided config. */
   def withTls(tls: TlsClientConfig): Router
 
+  def withAnnouncers(announcers: Seq[(Path, Announcer)]): Router
+
   /**
    * Initialize a router by instantiating a downstream router client
    * so that its upstream `servers` may be bound.
@@ -81,6 +83,7 @@ object Router {
     def protocol: ProtocolInitializer
     def params: Stack.Params
     def servers: Seq[Server.Initializer]
+    def announcers: Seq[(Path, Announcer)]
   }
 }
 
@@ -96,6 +99,9 @@ trait RouterConfig {
   var failFast: Option[Boolean] = None
   var timeoutMs: Option[Int] = None
   var dstPrefix: Option[String] = None
+
+  @JsonProperty("announcers")
+  var _announcers: Option[Seq[AnnouncerConfig]] = None
 
   @JsonProperty("label")
   var _label: Option[String] = None
@@ -175,8 +181,12 @@ trait RouterConfig {
   def router(params: Stack.Params): Router = {
     val prms = params ++ routerParams
     val param.Label(label) = prms[param.Label]
+    val announcers = _announcers.toSeq.flatten.map { announcer =>
+      announcer.prefix -> announcer.mk
+    }
     protocol.router.configured(prms)
       .serving(servers.map(_.mk(protocol, label)))
+      .withAnnouncers(announcers)
       .maybeTransform(client.flatMap(_.tls).map(tls => _.withTls(tls)))
   }